+++ embedded/lanyard/main/animation.h
@@ -4,7 +4,7 @@
 #define FADE 0
 #define RGB 1
 
-void animation_select(unsigned index);
+void animation_select(uint8_t index);
 void animation_start();
 void animation_enable();
 void animation_disable();
